(int)line_number

+++ interpret_lines.c
@@ -17,7 +17,7 @@ int handle_push_param(char **line, unsigned int line_number)
 		value = string_to_integer(line[1]);
 		if (value == NULL || *line[1] == '\0' || isspace(*line[1]))
 		{
-			print_monty_error(INVALID_INT, 1, line_number);
+			print_monty_error(INVALID_INT, 1, (int)line_number);
 			exit_code = EXIT_FAILURE;
 			free(value);
 		}
@@ -29,7 +29,7 @@ int handle_push_param(char **line, unsigned int line_number)
 	}
 	else
 	{
-		print_monty_error(INVALID_INT, 1, line_number);
+		print_monty_error(INVALID_INT, 1, (int)line_number);
 		exit_code = EXIT_FAILURE;
 	}
 	return (exit_code);
@@ -73,7 +73,7 @@ int process_line(stack_t **stack, char **line, unsigned int line_number)
 		}
 		if (!is_handeled_opcode)
 		{
-			print_monty_error(INVALID_OPCODE, 2, line_number, opcode);
+			print_monty_error(INVALID_OPCODE, 2, (int)line_number, opcode);
 			exit_code = EXIT_FAILURE;
 		}
 	}
